모달 백그라운드 설정 및 useState 적용
