+++ src/components/Modal/Fandom-k_Modal/module.css/Donation.module.css
@@ -5,7 +5,26 @@ input::-webkit-inner-spin-button {
 	-webkit-appearance: none;
 }
 
+/*모달 백그라운드 설정 및 버튼 누르기 전까지 숨기기*/
+#modalBackground {
+	width: 100%;
+	height: 100%;
+	position: fixed;
+	top: 0;
+	left: 0;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background: rgba(0, 0, 0, 0.5);
+	z-index: 100;
+}
+
+#modalBackground.hidden {
+	display: none;
+}
+
 /*연예인 후원하기 모달*/
+
 .donationContainer {
 	background-color: #181026;
 	width: 327px;
