+++ src/components/Modal/Fandom-k_Modal/module.css/Popup.module.css
@@ -1,4 +1,17 @@
 /*크레딧 부족 팝업 모달*/
+#modalBackground {
+	width: 100%;
+	height: 100%;
+	position: fixed;
+	top: 0;
+	left: 0;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background: rgba(0, 0, 0, 0.5);
+	z-index: 100;
+}
+
 .popupContainer {
 	width: 339px;
 	border-radius: 12px;
