+++ src/components/Modal/Fandom-k_Modal/module.css/Topup.module.css
@@ -1,4 +1,17 @@
 /* 크레딧 충전 모달 */
+#modalBackground {
+	width: 100%;
+	height: 100%;
+	position: fixed;
+	top: 0;
+	left: 0;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background: rgba(0, 0, 0, 0.5);
+	z-index: 100;
+}
+
 .container {
 	background-color: #181026;
 	width: 327px;
