+++ src/components/Modal/Fandom-k_Modal/module.css/Votes.module.css
@@ -1,4 +1,17 @@
 /* 연예인 투표 모달 */
+#modalBackground {
+	width: 100%;
+	height: 100%;
+	position: fixed;
+	top: 0;
+	left: 0;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background: rgba(0, 0, 0, 0.5);
+	z-index: 100;
+}
+
 .Container {
 	width: 525px;
 	background-color: #181026;
@@ -106,4 +119,13 @@
 .areaLine {
 	border-bottom: 1px solid #ffffff1a;
 	margin: 8px 0;
+}
+
+/*모달 open, close style*/
+.display-block {
+	display: block;
+}
+
+.display-none {
+	display: none;
 }
