+++ src/components/Modal/Modal.jsx
@@ -4,6 +4,30 @@ import Button from "../Button/Button";
 import Xbutton from "../Button/Xbutton";
 import "./Modal.css";
 
+//사용 방법 : 모달 컴포넌트를 사용할 컴포넌트에서 아래 코드를 작성
+/* App 컴포넌트에서 모달을 사용하는 경우 :
+import Modal from './Modal';
+import './App.css';
+import React, { useState } from 'react';
+
+const App = () => {
+	const [show, setShow] = useState(false);
+  
+	const handleOpen = () => setShow(true);
+	const handleClose = () => setShow(false);
+  
+	return (
+	  <div className="App">
+		<button onClick={handleOpen}>Open Modal</button>
+		<Modal show={show} handleClose={handleClose}>
+		  <p>This is a simple modal!</p>
+		</Modal>
+	  </div>
+	);
+  };
+  
+  export default App;*/
+
 function Modal({ show, icon, buttonAction, buttonName, onClose, children, title }) {
 	if (!show) {
 		document.body.style.removeProperty("overflow");
