+++ src/index.jsx
@@ -1,10 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Main from "./Main";
+//import Main from "./Main";
+import TestModal from "./TestModal";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-	<React.StrictMode>
-		<Main />,
-	</React.StrictMode>,
+	//<React.StrictMode>
+	//<Main />,
+	<TestModal />,
+	//</React.StrictMode>,
 );
